fix(stock_notification_service) Fixing the Kafka handle_user event handler
- Added an explicit event type check for calling the delete_user coroutine

+++ backend/stock_notification_service/stock_notification_service/stock_notification/kafka/kafka_handlers/user_handler.py
@@ -25,7 +25,7 @@ async def handle_user(key: str | None, value: dict):
                 user_update = UserUpdate.model_validate(data, strict=False)
                 await update_user(session, user_id, user_update)
 
-            else:
+            elif event_type == "user_deleted":
                 user_id = value.get("user_id")
                 await delete_user(session, user_id)
     except Exception as e:
